msm: subsystem_restart: Add support for ramdump notifications
When a subsystem crashes and userspace has enabled ramdumps
through the subsystem framework, a driver (which is not the
subsystem driver) might want to know about it so that it
can dump additional debug information. These drivers can't
use the existing "SUBSYS_AFTER_SHUTDOWN" notification since
they can't tell if the power down is due to a clean shutdown
or due to a crash. Also, having a notification that's always
sent out with a "ramdump enabled" flag would remove the
need for some of the drivers to maintain additional state.

So, add a new ramdump notification with an "enabled" flag
that's always sent out after a subsystem crashes to indicate
that the state of the system is conducive to dumping additional
debug information.

Change-Id: I6f5701f7003d1547e3dbf8cc7d9fbe09b230a7ed
Signed-off-by: Seemanta Dutta <seemanta@codeaurora.org>