power: pm8921-bms: remember the temperature when ocv was taken
The code looks up the percent charge based on ocv at the
current temperature. The ocv could have been recorded earlier
under different temperature conditions. Looking it up at the
current temperature is wrong.

Hence remember the temperature when the ocv was first seen. And
use that to lookup the ocv values.

Change-Id: Icdc707a9438adae5c8bb2d9b97af95e65c4bc2a1
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>