msm: msm_bus: Change initcall to subsys for bus devices
This change is required to enable early initialization of fabric
devices using device tree.
In the absence of this change, or if module_init/arch_init
are used, client registration is attempted before the
NoCs or fabrics are initialized. The registration fails
as the path cannot be calculated as the topology is built
after all the NoCs/Fabrics are initialized.

Change-Id: I79603af5d36e9afe1be25e4267cb01ba6a3c1d07
Signed-off-by: Gagan Mac <gmac@codeaurora.org>