8930: Use MPP line for USB ID detection
0n 8930 with PMIC 8917, USB ID line is not connected
to PMIC instead it is routed using MPP line. This
change adds support to configure MPP01 as input pin
and uses IRQ line for MPP01 for USB ID detection.

By default, this feature is disable as it is required
to have rework (i.e. external 30k pullup on MPP01 line)
with used hardware.

To enable this feature:
set usb_host_mode_pm8917=1 as boot param

CRs-Fixed: 438131
Change-Id: I7a3ef8cfc179b3504216673cf9e817b93e7a4b5d
Signed-off-by: Mayank Rana <mrana@codeaurora.org>