crypto: msm: Fix failure to process large packets
When performing a crypto operation of packet size larger
than 32KB -1 Bytes (max supported by BAM HW), the operation hangs.

The root casue was identified to be in one of the accumulators of
NDP_BAM not being given enough time to clear prior to when the EOT is
seen from the crypto core.

The workaround to the above issue in NDP_BAM is of two folds
- Re-order the data descriptor packet so that the last descriptor data
  (prior to the results dump data descriptor) is of max packet size.
- Decrease the max packet size per data descriptor to 16 KB

The above changes results in allowing enough time for the accumulators
to clear before the last results dump data descriptor is processed by
BAM.

Change-Id: Ic50eb50df57f0dd1d603b4db1f50386f2b79f712
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>